requirements.txt
Flask==0.11.1
locustio==0.7.5
numpy==1.11.2
Pillow==3.3.0
requests==2.11.1
tensorflow==0.11.0rc0
